resolve(ShopPage):button element resolved
