+++ constants/ShopPageSamples.ts
@@ -1,0 +1,52 @@
+interface Product {
+  title: string;
+  brand: string;
+  price: string;
+  description: string;
+  imageUrl: string;
+}
+
+const data: Product[] = [
+  {
+    title: "iPhone 14 Pro",
+    brand: "Apple",
+    price: "۱۰,۰۰۰ تومان",
+    description:
+      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
+    imageUrl: "./src/assets/images/iphone-14-pro.png",
+  },
+  {
+    title: "iPhone 14 Pro",
+    brand: "Apple",
+    price: "۱۰,۰۰۰ تومان",
+    description:
+      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
+    imageUrl: "./src/assets/images/iphone-14-pro.png",
+  },
+  {
+    title: "iPhone 14 Pro",
+    brand: "Apple",
+    price: "۱۰,۰۰۰ تومان",
+    description:
+      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
+    imageUrl: "./src/assets/images/iphone-14-pro.png",
+  },
+  {
+    title: "iPhone 14 Pro",
+    brand: "Apple",
+    price: "۱۰,۰۰۰ تومان",
+    description:
+      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
+    imageUrl: "./src/assets/images/iphone-14-pro.png",
+  },
+   {
+    title: "iPhone 14 Pro",
+    brand: "Apple",
+    price: "۱۰,۰۰۰ تومان",
+    description:
+      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
+    imageUrl: "./src/assets/images/iphone-14-pro.png",
+  },
+];
+
+export default data;
