+++ src/components/ShopPageProducts.tsx
@@ -1,54 +1,5 @@
 import ShopProductCard from "./ui/ShopProductCard";
-interface Product {
-  title: string;
-  brand: string;
-  price: string;
-  description: string;
-  imageUrl: string;
-}
-
-const data: Product[] = [
-  {
-    title: "iPhone 14 Pro",
-    brand: "Apple",
-    price: "۱۰,۰۰۰ تومان",
-    description:
-      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
-    imageUrl: "./src/assets/images/iphone-14-pro.png",
-  },
-  {
-    title: "iPhone 14 Pro",
-    brand: "Apple",
-    price: "۱۰,۰۰۰ تومان",
-    description:
-      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
-    imageUrl: "./src/assets/images/iphone-14-pro.png",
-  },
-  {
-    title: "iPhone 14 Pro",
-    brand: "Apple",
-    price: "۱۰,۰۰۰ تومان",
-    description:
-      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
-    imageUrl: "./src/assets/images/iphone-14-pro.png",
-  },
-  {
-    title: "iPhone 14 Pro",
-    brand: "Apple",
-    price: "۱۰,۰۰۰ تومان",
-    description:
-      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
-    imageUrl: "./src/assets/images/iphone-14-pro.png",
-  },
-   {
-    title: "iPhone 14 Pro",
-    brand: "Apple",
-    price: "۱۰,۰۰۰ تومان",
-    description:
-      "آیفون 14 پرو دارای صفحه نمایش 6.1 اینچی سوپر است نمایشگر Retina XDR ...",
-    imageUrl: "./src/assets/images/iphone-14-pro.png",
-  },
-];
+import data from "../../constants/ShopPageSamples";
 
 const ShopPage_Products = () => {
   return (
